feat(v2.1): split two values chart logic into a reusable component
Also, add a possibility of having just the primary value of the chart
